Fixed the about me section
I fixed the about me section by changing the p tag style into the article style. This made the CSS Grid more centered, although it does still show some issues with different screen resolutions.

The introductionary text at the start of the page is now an article. I could change this into a new div later on.

I added a new aboutme div and defined the style for the about me text in the css file.